feat: move RegisterOrLoginDTO to a shared location and update imports in auth and user services

+++ src/users/users.service.ts
@@ -1,7 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { UpdateUserDTO } from './dto/update-user.dto';
 import { UserSafeDTO } from './dto/user-safe.dto';
+import * as bcrypt from "bcrypt";
+import { RegisterOrLoginDTO } from '../dto/register-or-login.dto';
 
 @Injectable()
 export class UsersService {
@@ -40,7 +42,18 @@ export class UsersService {
     });
   }
 
-  async delete(id: number): Promise<UserSafeDTO> {
+  async delete(id: number, userData: RegisterOrLoginDTO): Promise<UserSafeDTO> {
+    const user = await this.prisma.user.findUniqueOrThrow({
+      where: {
+        id: id
+      }
+    });
+
+    const matched = await bcrypt.compare(userData.password, user.password);
+    if(!matched) {
+      throw new UnauthorizedException("로그인되어 있지 않습니다.");
+    }
+
     return this.prisma.user.delete({
       select: {
         id: true,
